requirements.txt
# Note: Install PyTorch separately (CUDA build recommended):
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# -------------------------
# Core runtime (V2 / src/app.py)
# -------------------------
python-dotenv>=1.0.0
numpy>=1.24.0
sounddevice>=0.4.6

openai>=1.40.0
rich>=13.7.0

# App launching
AppOpener>=1.7

# Web search
ddgs>=9.0.0

# -------------------------
# ASR
# -------------------------
faster-whisper>=1.0.0

# -------------------------
# TTS
# -------------------------
qwen-tts
# Optional (only needed if you switch to KokoroTTS)
kokoro
kittentts
# -------------------------
# Router (src/router.py) — Tier 1 only
# -------------------------
transformers>=4.30.0
huggingface_hub>=0.23.0

# -------------------------
# LangGraph Agent (Tier 2)
# -------------------------
langgraph
langchain-core
langchain-groq
langgraph-checkpoint-sqlite

# -------------------------
# Spotify
# -------------------------
# (uses AppOpener + webbrowser, no extra deps)

# -------------------------
# Optional / experiments
# -------------------------
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
google-genai
PySide6
opencv-python
Pillow
websockets
pyaudio

# -------------------------
# PC Controls
# -------------------------
pycaw>=20181226                # Volume control
screen-brightness-control>=0.23.0  # Brightness control
pyautogui>=0.9.54              # Screenshot, minimize all

# -------------------------
# Research Tools
# -------------------------
arxiv>=2.1.0                   # Academic paper search
trafilatura>=1.6.0             # Webpage summarization
youtube-search-python>=1.6.6   # YouTube search

# -------------------------
# Testing
# -------------------------
pytest>=7.0.0                  # Test framework
pytest-cov>=4.0.0              # Coverage reports
